Tests/LibWeb/Text/input/Editing/beforeinput-event.html

<script src="../include.js"></script>
<style>
    #input {
        width: 100px;
        height: 100px;
        border: 1px solid black;
    }
</style>
<div id="input" contenteditable="true"></div>
<script>
    test(() => {
        const input = document.getElementById("input");
        input.addEventListener("beforeinput", (e) => {
            if (e.data !== 'r' && e.data !== 'e') {
                e.preventDefault();
                return;
            }
            println(`beforeinput data=(${e.data}) intputType=(${e.inputType})`);
        });
        internals.sendText(input, "raebece");
        internals.commitText();
        println(`Text in input: ${input.textContent}`);
    });
</script>
