Tests/LibWeb/Text/expected/Editing/beforeinput-event.txt

beforeinput data=(r) intputType=(insertText)
beforeinput data=(e) intputType=(insertText)
beforeinput data=(e) intputType=(insertText)
beforeinput data=(e) intputType=(insertText)
Text in input: reee
